Revert "Combine test files"
This reverts commit f6151012f1b309b00582ff2701b39a49f811b022.
