+++ test/tests/tsdx-build.test.js
@@ -1,0 +1,48 @@
+/**
+ * @jest-environment node
+ */
+'use strict';
+
+const shell = require('shelljs');
+const util = require('../fixtures/util');
+
+shell.config.silent = false;
+
+const stageName = 'stage-build';
+
+describe('tsdx build', () => {
+  beforeAll(() => {
+    util.teardownStage(stageName);
+  });
+
+  it('should compile files into a dist directory', () => {
+    util.setupStageWithFixture(stageName, 'build-default');
+
+    const output = shell.exec('node ../dist/index.js build');
+
+    expect(shell.test('-f', 'dist/index.js')).toBeTruthy();
+    expect(
+      shell.test('-f', 'dist/build-default.cjs.development.js')
+    ).toBeTruthy();
+    expect(
+      shell.test('-f', 'dist/build-default.cjs.production.js')
+    ).toBeTruthy();
+    expect(
+      shell.test('-f', 'dist/build-default.es.production.js')
+    ).toBeTruthy();
+    expect(
+      shell.test('-f', 'dist/build-default.umd.development.js')
+    ).toBeTruthy();
+    expect(
+      shell.test('-f', 'dist/build-default.umd.development.js')
+    ).toBeTruthy();
+
+    expect(shell.test('-f', 'dist/index.d.ts')).toBeTruthy();
+
+    expect(output.code).toBe(0);
+  });
+
+  afterEach(() => {
+    util.teardownStage(stageName);
+  });
+});
